rearranged project greened up test
